update get user api

+++ users/src/main/kotlin/com/pinpoint/users/user/UserRepository.kt
@@ -2,4 +2,4 @@ package com.pinpoint.users.user
 
 import org.springframework.data.repository.CrudRepository
 
-interface UserRepository : CrudRepository<User, Int>
+interface UserRepository : CrudRepository<User, Long>
